refactor(tooltip): changed default closeOnClick to false
