+++ packages/components/tooltip/tsconfig.json
@@ -1,4 +1,4 @@
 {
   "extends": "../../../tsconfig.json",
-  "include": ["src", "tests", "index.ts"]
+  "include": ["src", "tests", "index.ts", "tests"]
 }
